Move `suggestedStaySummaryListOptions` to utils
This renames the `defaultMatchingInformationValues` utils file to
`matchingInformationUtils` and then moves the main logic of the
`suggestedStaySummaryListOptions` getter method to the utils file
